+++ lib/src/widgets/center_loading.dart
@@ -4,11 +4,13 @@ import 'package:flutter/cupertino.dart';
 class CenterLoading extends StatelessWidget {
   final Color? color;
   final bool isAnimating;
+  final double radius;
 
   const CenterLoading({
     super.key,
     this.color,
     this.isAnimating = true,
+    this.radius = 10,
   });
 
   @override
@@ -17,6 +19,7 @@ class CenterLoading extends StatelessWidget {
       child: CupertinoActivityIndicator(
         color: color,
         animating: isAnimating,
+        radius: radius,
       ),
     );
   }